+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- [PR#17](https://github.com/DmitryTsepelev/graphql-ruby-fragment_cache/pull/17) Properly build cache keys based on input arguments ([@DmitryTsepelev][])
+
 ## 0.1.2 (2020-04-24)
 
 - [PR#16](https://github.com/DmitryTsepelev/graphql-ruby-fragment_cache/pull/16) Railtie turns off caching in test environment ([@DmitryTsepelev][])